+++ sampledb/config.py
@@ -463,6 +463,8 @@ FEDERATION_UUID = None
 ALLOW_HTTP = False
 VALID_TIME_DELTA = 300
 
+ENABLE_BACKGROUND_TASKS = False
+
 # environment variables override these values
 use_environment_configuration(env_prefix='SAMPLEDB_')
 
@@ -505,6 +507,7 @@ for config_name in {
     'SHOW_OBJECT_TITLE',
     'HIDE_OBJECT_TYPE_AND_ID_ON_OBJECT_PAGE',
     'DISABLE_INLINE_EDIT',
+    'ENABLE_BACKGROUND_TASKS',
 }:
     value = globals().get(config_name)
     if isinstance(value, str):